requirements.txt
ddddocr==1.0.6
PyYAML==6.0
requests==2.28.1
selenium==4.4.3
